cleankedex/pyproject.toml
[tool.poetry]
name = "cleankedex"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "cleankedex", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.23"
pythondi = "^1.2.4"
fastapi = "^0.105.0"
py-automapper = "^1.2.3"
uvicorn = "^0.25.0"
psycopg2-binary = "^2.9.9"
fastapi-pagination = "^0.12.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
